# Conflicts:
#	src/main/java/kroryi/dagon/entity/Partner.java
#	src/main/java/kroryi/dagon/entity/PartnerApplication.java
#	src/main/java/kroryi/dagon/entity/ProductFacility.java
#	src/main/java/kroryi/dagon/entity/ProductFishSpecies.java
#	src/main/java/kroryi/dagon/entity/ProductFishingGear.java
#	src/main/java/kroryi/dagon/entity/Reservation.java

+++ src/main/java/kroryi/dagon/entity/ProductImg.java
@@ -18,7 +18,7 @@ public class ProductImg {
     private Long prodImgId;
 
     @Column(name = "prodid", nullable = false)
-    private Integer prodid;
+    private Long prodid;
 
     @Column(name = "prod_img_order", nullable = false)
     private Integer prodImgOrder;
